feat: switch logging to be asynchronous
This changes the backend of the Splunk longer to match more closely the
architecture of the built-in [Console
logger](https://github.com/elixir-lang/elixir/blob/master/lib/logger/lib/logger/backends/console.ex).

- If we have no background tasks when a message comes in, send it
immediately.
- If we do have a background task and have room in the queue, queue it up to
be sent later and return immediately.
- If we do not have room in the queue, then wait for the existing request to
finish, then send the entire queue.

This applies some backpressure if we're unable to send messages, which will
then cause the Logger backend itself to apply backpressure to the application.